# app/src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "User Management System",
    "description": "Create and Manage Users, Roles and Permissions",
    "contact": {
      "name": "Bibek Aryal"
    },
    "license": {
      "name": "Personal Use Only"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/ums",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/apps/app/{id}": {
      "get": {
        "tags": [
          "Apps Management"
        ],
        "summary": "Get app by ID",
        "description": "Retrieves a specific app by its unique ID",
        "operationId": "readApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - App Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      },
      "put": {
        "tags": [
          "Apps Management"
        ],
        "summary": "Update an app by ID",
        "description": "Updates an existing app identified by its unique ID",
        "operationId": "updateApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Details of the app to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "App updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Required Element Missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - App Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Apps Management"
        ],
        "summary": "Soft delete an app by ID",
        "description": "Marks an app as deleted without permanently removing it, identified by its unique ID",
        "operationId": "softDeleteApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app to soft delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App soft deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - App Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
    "/api/v1/apps/app": {
      "post": {
        "tags": [
          "Apps Management"
        ],
        "summary": "Create a new app",
        "description": "Creates a new app",
        "operationId": "createApp",
        "requestBody": {
          "description": "Details of the app to create",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "App created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Required Element Missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
    "/api/v1/apps/app/{id}/restore": {
      "patch": {
        "tags": [
          "Apps Management"
        ],
        "summary": "Restore a soft-deleted app by ID",
        "description": "Restores an app that was previously soft-deleted, identified by its unique ID",
        "operationId": "restoreApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app to restore",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App restored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - App Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
    "/api/v1/apps": {
      "get": {
        "tags": [
          "Apps Management"
        ],
        "summary": "Get all apps",
        "description": "Retrieves all apps available in the system",
        "operationId": "readApps",
        "responses": {
          "200": {
            "description": "Apps retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
    "/api/v1/apps/app/{id}/hard": {
      "delete": {
        "tags": [
          "Apps Management"
        ],
        "summary": "Hard delete an app by ID",
        "description": "Permanently removes an app identified by its unique ID",
        "operationId": "hardDeleteApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app to hard delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App hard deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - App Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppsRequest": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "AppsDto": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        }
      },
      "AppsResponse": {
        "type": "object",
        "properties": {
          "responseCrudInfo": {
            "$ref": "#/components/schemas/ResponseCrudInfo"
          },
          "responsePageInfo": {
            "$ref": "#/components/schemas/ResponsePageInfo"
          },
          "responseStatusInfo": {
            "$ref": "#/components/schemas/ResponseStatusInfo"
          },
          "apps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppsDto"
            }
          }
        }
      },
      "ResponseCrudInfo": {
        "type": "object",
        "properties": {
          "insertedRowsCount": {
            "type": "integer",
            "format": "int32"
          },
          "updatedRowsCount": {
            "type": "integer",
            "format": "int32"
          },
          "deletedRowsCount": {
            "type": "integer",
            "format": "int32"
          },
          "restoredRowsCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ResponsePageInfo": {
        "type": "object",
        "properties": {
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "perPage": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ResponseStatusInfo": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "errMsg": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Token": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      },
      "Basic": {
        "type": "http",
        "scheme": "Basic"
      }
    }
  }
}